Added deleteUser spec

+++ backend/src/modules/user/controllers/deleteUser.js
@@ -1,3 +1,4 @@
+const { isValidObjectId } = require('mongoose');
 const { userProvider } = require('../../../../db/providers');
 const { errorHandler } = require('../../../utils');
 
@@ -5,9 +6,16 @@ module.exports = async (req, res) => {
   try {
     const { userId } = req.params;
 
-    const deletedUser = await userProvider.deleteById(userId);
+    if (isValidObjectId(userId)) {
+      const deletedUser = await userProvider.deleteById(userId);
 
-    return res.status(200).json(deletedUser);
+      if (!deletedUser) {
+        return res.status(404).json({ message: 'User does not exist.' });
+      }
+
+      return res.status(200).json(deletedUser);
+    }
+    return res.status(400).json({ message: 'Provided userId is no valid.' });
   } catch (error) {
     return errorHandler(res, error);
   }
